all requests sorted

+++ DapperCRUDAPI/Models/Movie.cs
@@ -11,6 +11,6 @@ namespace DapperCRUDAPI.Models
         public string Genre { get; set; }
         public string Rating { get; set; }
         public string ReleaseDate { get; set; }
-        public double IMDbScore { get; set; }
+        public decimal IMDbScore { get; set; }
     }
 }
